lesson7_2.7_явное преобразование типов_float->int

+++ lesson7/2.go
@@ -10,6 +10,6 @@ func main() {
 	intVar := 10
 	floatVar := 10.2
 
-	fmt.Println(float64(intVar) + floatVar)
+	fmt.Println(intVar + int(floatVar))
 
 }
